Wrap oversized GSSAPI payloads correctly
The original [MS-NNS] specification incorrectly noted the max length of
negotiate stream packets as 0xFC00, when this was max length of the
pre-wrapped data. The original code also did not GSSAPI wrap in chunks.

Wrapping each chunk adds a header of up to 0x30, which has now been
corrected in the upstream specification.

https://lists.samba.org/archive/cifs-protocol/2018-December/003195.html

(Note: There may also need to be a correction on the read side.)